# Conflicts:
#	build.gradle
#	src/main/java/com/studycollaboproject/scope/service/PostService.java
#	src/main/java/com/studycollaboproject/scope/service/UserService.java

+++ src/main/java/com/studycollaboproject/scope/service/TeamService.java
@@ -1,8 +1,9 @@
 package com.studycollaboproject.scope.service;
 
 import com.studycollaboproject.scope.dto.MemberListResponseDto;
-import com.studycollaboproject.scope.exception.ErrorCode;
 import com.studycollaboproject.scope.exception.BadRequestException;
+import com.studycollaboproject.scope.exception.ErrorCode;
+import com.studycollaboproject.scope.exception.ForbiddenException;
 import com.studycollaboproject.scope.model.*;
 import com.studycollaboproject.scope.repository.ApplicantRepository;
 import com.studycollaboproject.scope.repository.TeamRepository;
@@ -22,14 +23,21 @@ public class TeamService {
     private final ApplicantRepository applicantRepository;
 
     @Transactional
-    public void acceptMember(Post post, User user, Boolean accept) {
+    public boolean acceptMember(Post post, User user, Boolean accept) {
         // [예외처리] 대기열에서 신청자의 정보를 찾을 수 없을 때
         Applicant applicant = applicantRepository.findByUserAndPost(user, post).orElseThrow(
                 () -> new BadRequestException(ErrorCode.NO_APPLICANT_ERROR)
         );
+        if (accept && post.getTotalMember() == post.getRecruitmentMember()) {
+            throw new BadRequestException(ErrorCode.NO_EXCEED_MEMBER_ERROR);
+        }
         //대기열에서 삭제
         applicant.deleteApply();
         applicantRepository.delete(applicant);
+
+        if (teamRepository.existsByPostIdAndUserSnsId(post.getId(), user.getSnsId())) {
+            throw new BadRequestException(ErrorCode.ALREADY_TEAM_ERROR);
+        }
 
         if (accept) {
             Team team = Team.builder()
@@ -37,8 +45,9 @@ public class TeamService {
                     .post(post)
                     .build();
             teamRepository.save(team);
+            return true;
         }
-
+        return false;
     }
 
     @Transactional
@@ -47,22 +56,21 @@ public class TeamService {
                 .stream().map(MemberListResponseDto::new).collect(Collectors.toCollection(ArrayList::new));
     }
 
-    public void memberResignation(User member, Post post) {
-        if (post.getProjectStatus().equals(ProjectStatus.PROJECT_STATUS_RECRUITMENT)) {
-            teamRepository.deleteByUserAndPost(member, post);
-        } else throw new BadRequestException(ErrorCode.NO_RECRUITMENT_ERROR);
-    }
-
-    public void memberSecession(User user, Post post) {
-
-        if (post.getProjectStatus().equals(ProjectStatus.PROJECT_STATUS_RECRUITMENT)) {
+    @Transactional
+    public void memberDelete(User user, Post post) {
+        if (post.getUser().getId().equals(user.getId())) {
+            throw new BadRequestException(ErrorCode.NOT_AVAILABLE_ACCESS);
+        } else if (post.getProjectStatus().equals(ProjectStatus.PROJECT_STATUS_RECRUITMENT)) {
+            Team team = teamRepository.findByUserAndPost(user, post).orElseThrow(
+                    () -> new ForbiddenException(ErrorCode.NO_TEAM_ERROR)
+            );
+            team.deleteTeam();
             teamRepository.deleteByUserAndPost(user, post);
         } else throw new BadRequestException(ErrorCode.NO_RECRUITMENT_ERROR);
-
     }
 
-    public boolean isMemeber(Post post, User user) {
-        return teamRepository.findByUserAndPost(user,post).isPresent();
+    public boolean isMember(Post post, User user) {
+        return teamRepository.findByUserAndPost(user, post).isPresent();
 
     }
 }
